refactor(schema): merge domain and interface directories
