+++ packages/specs/schema/src/components/mappers/default/anyMapper.ts
@@ -1,6 +1,6 @@
 import {JsonLazyRef} from "../../../domain/JsonLazyRef.js";
 import {JsonSchema} from "../../../domain/JsonSchema.js";
-import {JsonSchemaOptions} from "../../../interfaces/JsonSchemaOptions.js";
+import {JsonSchemaOptions} from "../../../domain/JsonSchemaOptions.js";
 import {defineSchemaMapper, execMapper, execOneOfMapper} from "../../../registries/JsonSchemaMapperContainer.js";
 import {toRef} from "../../../utils/ref.js";
