+++ packages/specs/schema/src/components/mappers/default/classMapper.ts
@@ -1,7 +1,7 @@
 import {getValue, setValue} from "@tsed/core";
 
 import {JsonSchema} from "../../../domain/index.js";
-import {JsonSchemaOptions} from "../../../interfaces/JsonSchemaOptions.js";
+import {JsonSchemaOptions} from "../../../domain/JsonSchemaOptions.js";
 import {getJsonEntityStore} from "../../../registries/JsonEntitiesContainer.js";
 import {defineSchemaMapper, execMapper} from "../../../registries/JsonSchemaMapperContainer.js";
 import {createRef, createRefName} from "../../../utils/ref.js";
