+++ packages/specs/schema/src/components/mappers/default/inheritedClassMapper.ts
@@ -1,6 +1,6 @@
 import {classOf, deepMerge} from "@tsed/core";
 
-import {JsonSchemaOptions} from "../../../interfaces/JsonSchemaOptions.js";
+import {JsonSchemaOptions} from "../../../domain/JsonSchemaOptions.js";
 import {defineSchemaMapper, execMapper} from "../../../registries/JsonSchemaMapperContainer.js";
 import {getInheritedStores} from "../../../utils/getInheritedStores.js";
