+++ packages/specs/schema/src/components/mappers/default/requiredMapper.ts
@@ -1,8 +1,8 @@
 import {uniq} from "@tsed/core";
 
 import type {JsonSchema} from "../../../domain/JsonSchema.js";
+import type {JsonSchemaOptions} from "../../../domain/JsonSchemaOptions.js";
 import {alterRequiredGroups} from "../../../hooks/alterRequiredGroups.js";
-import type {JsonSchemaOptions} from "../../../interfaces/JsonSchemaOptions.js";
 import {defineSchemaMapper} from "../../../registries/JsonSchemaMapperContainer.js";
 
 function mapRequiredProps(obj: any, schema: JsonSchema, options: JsonSchemaOptions = {}) {
