+++ packages/specs/schema/src/components/mappers/open-spec/operationInParameterMapper.ts
@@ -4,7 +4,7 @@ import type {JSONSchema6} from "json-schema";
 
 import {JsonParameter} from "../../../domain/JsonParameter.js";
 import {JsonParameterTypes} from "../../../domain/JsonParameterTypes.js";
-import {JsonSchemaOptions} from "../../../interfaces/JsonSchemaOptions.js";
+import {JsonSchemaOptions} from "../../../domain/JsonSchemaOptions.js";
 import {defineSchemaMapper, execMapper, hasMapper} from "../../../registries/JsonSchemaMapperContainer.js";
 
 export type JsonParameterOptions = JsonSchemaOptions & {
