+++ packages/specs/schema/src/components/mappers/open-spec/operationInParametersMapper.ts
@@ -1,5 +1,5 @@
 import {JsonParameter} from "../../../domain/JsonParameter.js";
-import {JsonSchemaOptions} from "../../../interfaces/JsonSchemaOptions.js";
+import {JsonSchemaOptions} from "../../../domain/JsonSchemaOptions.js";
 import {defineSchemaMapper, execMapper} from "../../../registries/JsonSchemaMapperContainer.js";
 
 export function operationInParametersMapper(parameters: JsonParameter[], options: JsonSchemaOptions) {
