+++ packages/specs/schema/src/components/mappers/open-spec/operationMediaMapper.ts
@@ -1,5 +1,5 @@
 import {JsonMedia} from "../../../domain/JsonMedia.js";
-import {JsonSchemaOptions} from "../../../interfaces/JsonSchemaOptions.js";
+import {JsonSchemaOptions} from "../../../domain/JsonSchemaOptions.js";
 import {defineSchemaMapper, execMapper} from "../../../registries/JsonSchemaMapperContainer.js";
 
 export function operationMediaMapper(jsonMedia: JsonMedia, options: JsonSchemaOptions) {
