+++ packages/specs/schema/src/components/mappers/open-spec/operationRequestBodyMapper.ts
@@ -1,7 +1,7 @@
 import {JsonParameter} from "../../../domain/JsonParameter.js";
 import {JsonRequestBody} from "../../../domain/JsonRequestBody.js";
 import {JsonSchema} from "../../../domain/JsonSchema.js";
-import {JsonSchemaOptions} from "../../../interfaces/JsonSchemaOptions.js";
+import {JsonSchemaOptions} from "../../../domain/JsonSchemaOptions.js";
 import {defineSchemaMapper, execMapper} from "../../../registries/JsonSchemaMapperContainer.js";
 
 function buildSchemaFromBodyParameters(parameters: JsonParameter[], options: JsonSchemaOptions) {
