+++ packages/specs/schema/src/decorators/operations/location.ts
@@ -1,4 +1,4 @@
-import {JsonHeader} from "../../interfaces/JsonOpenSpec.js";
+import {JsonHeader} from "../../domain/JsonOpenSpec.js";
 import {Returns} from "./returns.js";
 
 /**
