+++ packages/specs/schema/src/domain/IgnoreCallback.ts
