+++ packages/specs/schema/src/domain/JsonHookContext.ts
