+++ packages/specs/schema/src/domain/JsonOpenSpec.ts
