+++ packages/specs/schema/src/hooks/alterRequiredGroups.ts
@@ -1,5 +1,5 @@
 import type {JsonSchema} from "../domain/JsonSchema.js";
-import type {JsonSchemaOptions} from "../interfaces/JsonSchemaOptions.js";
+import type {JsonSchemaOptions} from "../domain/JsonSchemaOptions.js";
 /**
  * @ignore
  */
