+++ packages/specs/schema/src/index.ts
@@ -127,14 +127,17 @@ export * from "./decorators/operations/tags.js";
 export * from "./decorators/operations/view.js";
 export * from "./domain/DecoratorContext.js";
 export * from "./domain/Generics.js";
+export * from "./domain/IgnoreCallback.js";
 export * from "./domain/index.js";
 export * from "./domain/JsonAliasMap.js";
 export * from "./domain/JsonDiscriminator.js";
 export * from "./domain/JsonEntityStore.js";
 export * from "./domain/JsonFormatTypes.js";
+export * from "./domain/JsonHookContext.js";
 export * from "./domain/JsonLazyRef.js";
 export * from "./domain/JsonMap.js";
 export * from "./domain/JsonMedia.js";
+export * from "./domain/JsonOpenSpec.js";
 export * from "./domain/JsonOperation.js";
 export * from "./domain/JsonOperationPathsMap.js";
 export * from "./domain/JsonOperationRoute.js";
@@ -143,6 +146,7 @@ export * from "./domain/JsonParameterTypes.js";
 export * from "./domain/JsonRequestBody.js";
 export * from "./domain/JsonResponse.js";
 export * from "./domain/JsonSchema.js";
+export * from "./domain/JsonSchemaOptions.js";
 export * from "./domain/SpecTypes.js";
 export * from "./domain/types.js";
 export * from "./fn/allOf.js";
@@ -179,10 +183,6 @@ export * from "./fn/url.js";
 export * from "./hooks/alterIgnore.js";
 export * from "./hooks/alterOneOf.js";
 export * from "./hooks/alterRequiredGroups.js";
-export * from "./interfaces/IgnoreCallback.js";
-export * from "./interfaces/JsonHookContext.js";
-export * from "./interfaces/JsonOpenSpec.js";
-export * from "./interfaces/JsonSchemaOptions.js";
 export * from "./registries/enumRegistry.js";
 export * from "./registries/JsonEntitiesContainer.js";
 export * from "./registries/JsonSchemaMapperContainer.js";
