+++ packages/specs/schema/src/utils/mapHeaders.ts
@@ -1,6 +1,6 @@
 import {cleanObject} from "@tsed/core";
 
-import {JsonHeader, JsonHeaders} from "../interfaces/JsonOpenSpec.js";
+import {JsonHeader, JsonHeaders} from "../domain/JsonOpenSpec.js";
 
 /**
  * Map input header to a standard open spec header
