+++ packages/specs/schema/src/utils/operationIdFormatter.ts
@@ -1,6 +1,6 @@
 import {camelCase} from "change-case";
 
-import {JsonSchemaOptions} from "../interfaces/JsonSchemaOptions.js";
+import {JsonSchemaOptions} from "../domain/JsonSchemaOptions.js";
 
 const DEFAULT_PATTERN = "%c.%m";
